third_party/blink/web_tests/platform/generic/fast/events/touch/multi-touch-inside-iframes-expected.txt

Test multi-touch is sent to the same iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

First touch is on iframe1.
PASS document.elementFromPoint(101, 101).id is "iframe1"
PASS ev.type is "touchstart"
PASS touchreceiver is "iframe1"
PASS ev.changedTouches.length is 1
PASS ev.touches.length is 1
PASS ev.targetTouches.length is 1
PASS ev.touches[0].pageX is 50
PASS ev.touches[0].pageY is 50
PASS ev.touches[0].identifier is 0
PASS ev.touches[0].target.ownerDocument.title is "iframe1"
PASS ev.changedTouches[0].target.ownerDocument.title is "iframe1"
PASS ev.targetTouches[0].target.ownerDocument.title is "iframe1"


Second touch is on iframe2, nothing should happen.
PASS document.elementFromPoint(251, 101).id is "iframe2"


Moves the first touch outside iframe1.
PASS ev.type is "touchmove"
PASS touchreceiver is "iframe1"
PASS ev.changedTouches.length is 1
PASS ev.touches.length is 2
PASS ev.targetTouches.length is 1
PASS ev.touches[0].pageX is 150
PASS ev.touches[0].pageY is 150
PASS ev.touches[0].identifier is 0
PASS ev.touches[1].identifier is 1
PASS ev.changedTouches[0].identifier is 0
PASS ev.touches[0].target.ownerDocument.title is "iframe1"
PASS ev.changedTouches[0].target.ownerDocument.title is "iframe1"
PASS ev.targetTouches[0].target.ownerDocument.title is "iframe1"
PASS ev.touches[1].target.nodeName is "#document"
PASS ev.touches[1].target.title is "iframe1"


Release the first touch.
PASS ev.type is "touchend"
PASS touchreceiver is "iframe1"
PASS ev.changedTouches.length is 1
PASS ev.touches.length is 1
PASS ev.targetTouches.length is 0
PASS ev.changedTouches[0].target.ownerDocument.title is "iframe1"
PASS ev.changedTouches[0].identifier is 0
PASS ev.touches[0].target.nodeName is "#document"
PASS ev.touches[0].target.title is "iframe1"
PASS ev.touches[0].identifier is 1


Third touch is on iframe2, nothing should happen.
PASS document.elementFromPoint(261, 101).id is "iframe2"


Release all touches on iframe2, and touch iframe2 again.
PASS ev.type is "touchstart"
PASS touchreceiver is "iframe2"
PASS ev.changedTouches.length is 1
PASS ev.touches.length is 1
PASS ev.targetTouches.length is 1
PASS ev.touches[0].pageX is 50
PASS ev.touches[0].pageY is 50
PASS ev.touches[0].target.ownerDocument.title is "iframe2"
PASS ev.changedTouches[0].target.ownerDocument.title is "iframe2"
PASS ev.targetTouches[0].target.ownerDocument.title is "iframe2"


Touch iframe1, nothing shoud happen.


Release all touches.
PASS ev.type is "touchend"
PASS touchreceiver is "iframe2"
PASS ev.changedTouches.length is 1
PASS ev.touches.length is 0
PASS ev.targetTouches.length is 0
PASS ev.changedTouches[0].target.ownerDocument.title is "iframe2"
PASS successfullyParsed is true

TEST COMPLETE
